update(redirect): fix fmt and sep file for norm
