+++ Makefile
@@ -87,6 +87,7 @@ EXPAND_DIR	:=	expand
 SRCS		+=	$(EXEC_DIR)/$(EXPAND_DIR)/concat_tokens.c  \
 				$(EXEC_DIR)/$(EXPAND_DIR)/dirp.c\
 				$(EXEC_DIR)/$(EXPAND_DIR)/expand_for_heredoc.c\
+				$(EXEC_DIR)/$(EXPAND_DIR)/expand_for_redirect.c\
 				$(EXEC_DIR)/$(EXPAND_DIR)/expand_parameter.c\
 				$(EXEC_DIR)/$(EXPAND_DIR)/expand_var_in_heredoc.c\
 				$(EXEC_DIR)/$(EXPAND_DIR)/expansion.c  \
