+++ srcs/exec/redirect_connect_to_proc.c
@@ -36,7 +36,8 @@ t_result	connect_redirect_to_proc(t_ast *self_node)
 {
 	if (self_node->proc_fd[IN] != IN_FD_INIT)
 	{
-		if (connect_redirect_in_to_proc(&self_node->prev_fd, self_node->proc_fd[IN]) == PROCESS_ERROR)
+		if (connect_redirect_in_to_proc(\
+			&self_node->prev_fd, self_node->proc_fd[IN]) == PROCESS_ERROR)
 		{
 			close_proc_out_fd(self_node->proc_fd[OUT]);
 			return (PROCESS_ERROR);
@@ -44,7 +45,8 @@ t_result	connect_redirect_to_proc(t_ast *self_node)
 	}
 	if (self_node->proc_fd[OUT] != OUT_FD_INIT)
 	{
-		if (connect_redirect_out_to_proc(self_node->proc_fd[OUT]) == PROCESS_ERROR)
+		if (connect_redirect_out_to_proc(\
+			self_node->proc_fd[OUT]) == PROCESS_ERROR)
 		{
 			close_proc_in_fd(self_node->proc_fd[IN]);
 			return (PROCESS_ERROR);
